Refactored KTypeProbe
Decided to remove the templated default funcions - whilst the code certainly is tighter it seems fairly non idiomatic C++. For the STM32 we're not that concerned about every byte being saved
in the final binary. Added an abstract superclass so that the PID controller can be decoupled from the particular sensor it is controlling.

+++ include/CommandProcessor.h
@@ -17,7 +17,7 @@ class CommandProcessor {
 
     private:
         void jumpToBootloader();
-        void setTemperature(double temp);
+        void setTemperature(int temp);
         void processCommand();
 
     public:
